catch possible errors, add customer number, fix ui

+++ Atlas/Pages/2ndPageAddDel.xaml.cs
@@ -2,6 +2,7 @@
 using Microsoft.EntityFrameworkCore;
 using System;
 using System.Collections.Generic;
+using System.Globalization;
 using System.Linq;
 using System.Text;
 using System.Windows;
@@ -21,6 +22,7 @@ namespace Atlas.Pages
     /// </summary>
     public partial class _2ndPageAddDel : Page
     {
+        private static float Price;
         public _2ndPageAddDel()
         {
             InitializeComponent();
@@ -57,63 +59,71 @@ namespace Atlas.Pages
 
             using (DataContext context = new DataContext())
             {
-                var finOrder = context.Orders.FromSqlRaw("Select * From Orders").ToList();
-
-                int custQuantity = 0;
-                float custTotal = 0;
-
-                foreach (var eachorder in finOrder)
+                if (initial_Order.HasItems)
                 {
-                    var finalID = int.Parse(eachorder.ProductID.ToString());
-                    var finalQuan = int.Parse(eachorder.Quantity.ToString());
-                    var finalpri = float.Parse(eachorder.Price.ToString());
-                    var finaltot = float.Parse(eachorder.Total.ToString());
+                    var finOrder = context.Orders.FromSqlRaw("Select * From Orders").ToList();
 
-                    context.Orderitems.Add(new CSOrderitems()
+                    int custQuantity = 0;
+                    float custTotal = 0;
+
+                    foreach (var eachorder in finOrder)
+                    {
+                        var finalID = int.Parse(eachorder.ProductID.ToString());
+                        var finalQuan = int.Parse(eachorder.Quantity.ToString());
+                        var finalpri = float.Parse(eachorder.Price.ToString());
+                        var finaltot = float.Parse(eachorder.Total.ToString());
+
+                        context.Orderitems.Add(new CSOrderitems()
+                        {
+                            TrackingNumber = TrackingNum,
+                            ProductID = finalID,
+                            Quantity = finalQuan,
+                            UnitPrice = finalpri,
+                            TotPrice = finaltot
+                        });
+
+                        custQuantity += finalQuan;
+                        custTotal += finaltot;
+
+
+                    }
+                    foreach (var item in finOrder)
+                    {
+                        context.Orders.Remove(item);
+                    }
+                    DateTime date = DateTime.Now;
+                    CultureInfo ci = CultureInfo.InvariantCulture;
+
+                    var orderdate = date.ToString("yyyy-MM-dd HH:mm:ss", ci);
+                    var customerid = AddDelivery.selectedCus.ID;
+                    var custaddress = AddDelivery.selectedCus.Address;
+
+
+                    context.Deliveries.Add(new CSDelivery()
                     {
                         TrackingNumber = TrackingNum,
-                        ProductID = finalID,
-                        Quantity = finalQuan,
-                        UnitPrice = finalpri,
-                        TotPrice = finaltot
+                        CustomerID = customerid,
+                        Address = custaddress,
+                        Amount = custTotal,
+                        Quantity = custQuantity,
+                        OrderDate = orderdate
+
                     });
 
-                    custQuantity += finalQuan;
-                    custTotal += finaltot; 
-                   
-                    
+                    context.SaveChanges();
+                    MessageBox.Show("Done!");
+                    Read();
                 }
-                foreach (var item in finOrder)
-                {
-                    context.Orders.Remove(item);
-                }
-
-
-                var customerid = AddDelivery.selectedCus.ID;
-                var custaddress = AddDelivery.selectedCus.Address;
-
-
-                context.Deliveries.Add(new CSDelivery()
-                {
-                   TrackingNumber = TrackingNum,
-                   CustomerID = customerid,
-                   Address = custaddress,
-                   Amount = custTotal,
-                   Quantity = custQuantity
-                });
-
-                context.SaveChanges();
-                
-
-               
-                MessageBox.Show("Done!");
-                Read();
+                else
+                    MessageBox.Show("No products to add!");
             }
+            
             
         }
         private void btn_Order_Click(object sender, RoutedEventArgs e)
         {
             orderBtn(sender);
+            totalamount.Text = Convert.ToString(  Price * float.Parse(quantityValue.Text));
         }
 
         private void orderBtn(object sender)
@@ -123,82 +133,104 @@ namespace Atlas.Pages
 
             if (targetbutton != null && targetbutton.Name == "btn_Order")
             {
-
-
-                using (DataContext context = new DataContext())
+                if (inventory_list.SelectedItems.Count > 0)
                 {
-                    var db = new DataContext();
-
-                    CSProduct selProduct = (CSProduct)inventory_list.SelectedItems[0];
-                    var quantityval = int.Parse(quantityValue.Text);
-                    var uprice = float.Parse(selProduct.Price.ToString());
-                    var productnameval = selProduct.ProductName.ToString();
-                    var prodid = int.Parse(selProduct.ID.ToString());
-
-                    var iniTotal = quantityval * uprice;
-
-                    //CSCustomer selCustomer = (CSCustomer)customer_list.SelectedItems[0];
-                   
-                    var id = selProduct.ID;
-
-                    if (context.Orders.Any(e => e.ProductID == id))
+                    using (DataContext context = new DataContext())
                     {
-
-                        var insOrder = context.Orders.Single(b => b.ProductID == id);
-
-                        var requantityval = insOrder.Quantity + quantityval;
-
-                        insOrder.Quantity = requantityval;
-
-                        insOrder.Total = requantityval * insOrder.Price;
-
-                        MessageBox.Show("Hello");
                         
-                        context.Orders.Update(insOrder);
-                        selProduct.Stocks = selProduct.Stocks - quantityval;
-                        
-                        context.Products.Update(selProduct);
 
-                        context.SaveChanges();
-                    }
-                    else
-                    {
-                        context.Orders.Add(new Orderlist()
+                        var db = new DataContext();
+
+                        CSProduct selProduct = (CSProduct)inventory_list.SelectedItems[0];
+                        var quantityval = int.Parse(quantityValue.Text);
+                        var uprice = float.Parse(selProduct.Price.ToString());
+                        var productnameval = selProduct.ProductName.ToString();
+                        var prodid = int.Parse(selProduct.ID.ToString());
+
+                        var iniTotal = quantityval * uprice;
+                        Price = selProduct.Price;
+
+                        //CSCustomer selCustomer = (CSCustomer)customer_list.SelectedItems[0];
+
+                        var id = selProduct.ID;
+
+                        if (selProduct.Stocks > int.Parse(quantityValue.Text))
                         {
-                            ProductID = prodid,
-                            ProductName = productnameval,
-                            Quantity = quantityval,
-                            Price = uprice,
-                            Total = iniTotal
-                        });
-                        MessageBox.Show("bye");
-                        selProduct.Stocks = selProduct.Stocks - quantityval;
+                            
+                            if (context.Orders.Any(e => e.ProductID == id))
+                            {
 
-                        context.Products.Update(selProduct);
-                        context.SaveChanges();
-               
+                                var insOrder = context.Orders.Single(b => b.ProductID == id);
 
-                    }
+                                var requantityval = insOrder.Quantity + quantityval;
+
+                                insOrder.Quantity = requantityval;
+
+                                insOrder.Total = requantityval * insOrder.Price;
+
+                                
+                                MessageBox.Show("Hello");
+
+                                context.Orders.Update(insOrder);
+                                selProduct.Stocks = selProduct.Stocks - quantityval;
+
+                                context.Products.Update(selProduct);
+
+                                context.SaveChanges();
+
+                                
+
+
+                            }
+                            else
+                            {
+
+                                context.Orders.Add(new Orderlist()
+                                {
+                                    ProductID = prodid,
+                                    ProductName = productnameval,
+                                    Quantity = quantityval,
+                                    Price = uprice,
+                                    Total = iniTotal
+
+                                });
+                                MessageBox.Show("bye");
+                                selProduct.Stocks = selProduct.Stocks - quantityval;
+                                
+                                context.Products.Update(selProduct);
+                                context.SaveChanges();
+
+                                
+                                
+
+                            }
                             
 
-                    Read();
+                            Read();
+                        }
+                        else
+                            MessageBox.Show("Stocks is lower than the quantity required");
+
+                       
+                    }
+
+                    //Orderlist order = new Orderlist
+                    //{
+                    //    ProductName = productname,
+                    //    Quantity = quantity,
+                    //    Price = uprice,
+                    //    Total = iniTotal
+                    //};
 
 
 
+
+                    //MessageBox.Show(selCustomer.Address.ToString());
                 }
+                else
+                    MessageBox.Show("Please select a Product!");
+            
 
-                //Orderlist order = new Orderlist
-                //{
-                //    ProductName = productname,
-                //    Quantity = quantity,
-                //    Price = uprice,
-                //    Total = iniTotal
-                //};
-
-
-
-
-                //MessageBox.Show(selCustomer.Address.ToString());
             }
         }
         public void Read()
